games/requirements.txt
fastapi[standard]>=0.119.1
jinja2>=3.1.6
redis[hiredis]>=5.2.1
typer>=0.15.2
pydantic-settings[yaml]>=2.10.1
python-multipart>=0.0.20
sqlalchemy>=2.0.44
asyncpg>=0.30.0
jwt>=1.4.0
pyjwt[crypto]>=2.10.1
bcrypt>=5.0.0
pyopenssl>=25.3.0
cryptography>=46.0.3
yookassa>=3.9.0
var-dump>=1.2
yookassa-payout>=2.7.0
yoomoney>=0.1.2